requirements.txt
asttokens==2.4.1
backcall==0.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.1.1
crc32c==2.3.post0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
fonttools==4.44.3
fsspec==2023.12.1
gdown==4.7.1
huggingface-hub==0.19.4
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lightning-utilities==0.10.0
matplotlib==3.7.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.24.4
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
positional-encodings==6.0.1
prompt-toolkit==3.0.41
protobuf==3.20.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tfrecord==1.14.4
threadpoolctl==3.2.0
timm==0.9.12
torch==1.12.1+cu102
torchaudio==0.12.1+cu102
torchmetrics==1.2.0
torchvision==0.13.1+cu102
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
zipp==3.17.0
